Deploy-ML/requirements.txt
Flask
tensorflow
numpy
pandas
scikit-surprise
scikit-learn
gunicorn
google-cloud-firestore